# Remove chunker.Reset() and chunker pool