This removes pools.go and makes the Chunker use a simpler pool for it's buffer. Also make `chunker.Reset()` private (since we are not reusing chunker instances anymore).