@fw42 2015-05-04
As discussed in https://github.com/restic/restic/pull/166#discussion_r29563676, I would like `chunker.New()` to return an error if the given polynomial is not irreducible. Can be a follow-up PR though.