@fd0 2015-05-04
Interesting benchmarks:

    % benchcmp /tmp/{before,after}
    benchmark                         old ns/op      new ns/op      delta
    [...]
    BenchmarkArchiveDirectory         1863372190     1052511209     -43.52%
    BenchmarkLoadTree                 2108148        1419818        -32.65%
    BenchmarkChunkerWithSHA256        148980904      150942129      +1.32%
    BenchmarkChunkerWithMD5           94357514       95597196       +1.31%
    BenchmarkChunker                  74275711       75383805       +1.49%
    BenchmarkNewChunker               74945          607            -99.19%
    [...]

    benchmark                         old bytes     new bytes     delta
    [...]
    BenchmarkChunkEncrypt             64694         12556         -80.59%
    BenchmarkChunkEncryptParallel     55072         2940          -94.66%
    BenchmarkArchiveDirectory         130384200     130690536     +0.23%
    BenchmarkLoadTree                 165816        115941        -30.08%
    BenchmarkChunkerWithSHA256        581           828           +42.51%
    BenchmarkChunkerWithMD5           434           678           +56.22%
    BenchmarkChunker                  338           574           +69.82%
    BenchmarkNewChunker               262512        336           -99.87%
    [...]

So we didn't break anything serious. :+1: